TASK component from 2200 to 700 lines. created components and hooks ,categorized them in folders
