+++ client/src/pages/Tasks/components/detail/TaskDetail.js
@@ -5,24 +5,24 @@ import { GoClockFill, GoGoal } from "react-icons/go";
 import { MdDateRange } from "react-icons/md";
 import { RiTimerLine } from "react-icons/ri";
 import { format } from "date-fns";
-import Loader from "../../utlis/Loader";
-import { Timer } from "../../utlis/Timer";
+import Loader from "../../../../utlis/Loader";
+import { Timer } from "../../../../utlis/Timer";
 
 import { FaRegUser } from "react-icons/fa";
 import { FaEdit } from "react-icons/fa";
 import { AiFillDelete } from "react-icons/ai";
 import { IoClose } from "react-icons/io5";
-import { style } from "../../utlis/CommonStyle";
+import { style } from "../../../../utlis/CommonStyle";
 import Swal from "sweetalert2";
 import { MdCheckCircle } from "react-icons/md";
-import AddTaskModal from "../../components/Tasks/AddTaskModal";
+import AddTaskModal from "../../../../components/Tasks/AddTaskModal";
 import { DragDropContext, Droppable, Draggable } from "react-beautiful-dnd";
 import { IoCloseCircleOutline } from "react-icons/io5";
 import { RiLoaderFill } from "react-icons/ri";
 import { BiBellPlus } from "react-icons/bi";
 import { BiSolidBellPlus } from "react-icons/bi";
 
-import Reminder from "../../utlis/Reminder";
+import Reminder from "../../../../utlis/Reminder";
 import { RxPerson } from "react-icons/rx";
 import { useSelector } from "react-redux";
 import { FiClock } from "react-icons/fi";
