+++ client/src/pages/Tasks/components/detail/TaskDetailComments.js
@@ -3,13 +3,13 @@ import { IoClose } from "react-icons/io5";
 
 import toast from "react-hot-toast";
 import axios from "axios";
-import { style } from "../../utlis/CommonStyle";
+import { style } from "../../../../utlis/CommonStyle";
 import { BiLoaderCircle } from "react-icons/bi";
 import EmojiPicker from "emoji-picker-react";
 import { BsEmojiSmile } from "react-icons/bs";
 import { format } from "date-fns";
 import { AiFillLike, AiOutlineLike } from "react-icons/ai";
-import Loader from "../../utlis/Loader";
+import Loader from "../../../../utlis/Loader";
 import { GoDotFill } from "react-icons/go";
 
 import {
